increasing range of vars

+++ 0x13-bit_manipulation/5-flip_bits.c
@@ -10,7 +10,7 @@
 
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int diff, bits = 0;
+	unsigned long int diff, bits = 0;
 
 	for (diff = n ^ m; diff; diff = diff >> 1)
 		bits += diff & 1;
